fix: rainbow logo export moved to icons/src/index.js
