+++ packages/auth/docs/stories/emailPasswordSignUpForm.story.js
@@ -5,7 +5,7 @@ import app from '../../../../firebase';
 import WhenAuthenticated from '../../src/components/WhenAuthenticated';
 import WhenNoAuthenticated from '../../src/components/WhenNoAuthenticated';
 import EmailPasswordSignUpForm from '../../src/components/EmailPasswordSignUpForm';
-import RainbowLogo from './icons/rainbowLogo';
+import { RainbowLogo } from '../../../icons/src';
 
 export default {
     title: 'Auth/Stories',
