+++ packages/icons/src/index.js
@@ -1,10 +1,11 @@
 export { default as Close } from './close';
 export { default as Done } from './done';
 export { default as Error } from './error';
-export { default as Warning } from './warning';
-export { default as HomeBorder } from './homeBorder';
-export { default as HomeFilled } from './homeFilled';
 export { default as ExportBorder } from './exportBorder';
 export { default as ExportFilled } from './exportFilled';
+export { default as HomeBorder } from './homeBorder';
+export { default as HomeFilled } from './homeFilled';
 export { default as MoneyBorder } from './moneyBorder';
 export { default as MoneyFilled } from './moneyFilled';
+export { default as RainbowLogo } from './rainbowLogo';
+export { default as Warning } from './warning';
