+++ packages/icons/src/rainbowLogo.js
